% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_data}
\alias{read_data}
\title{Read raster data}
\usage{
read_data(
  filename = system.file("tif/L7_ETMs.tif", package = "stars"),
  crop = get_bb()
)
}
\arguments{
\item{filename}{chr the file to read}

\item{crop}{NULL or a bbox for cropping}
}
\value{
stars object
}
\description{
Read raster data
}
