% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starscrop-package.R
\docType{package}
\name{starscrop}
\alias{starscrop-package}
\alias{starscrop}
\title{starscrop}
\description{
Demostrate issue with stars::crop() in a package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/starscrop}
}

}
\keyword{internal}
